server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-openai
pypdf
python-dotenv 